requirements.txt
sqlparse==0.3.0
gitpython==2.1.11
pyyaml==5.1
